00_Система/Инструменты/requirements.txt
# AGrav Smart Parsing & Translation Toolkit Requirements
# Install via: pip install -r requirements.txt

# PDF Processing & Batching
pymupdf>=1.24.0

# "Ultimate" Layout Recovery & Table Parsing
docling>=1.0.0
docling-core>=1.0.0

# OCR (Local engines)
# Note: RapidOCR is used as a fallback/enhancement in Docling
rapidocr-onnxruntime>=1.3.0

# AI APIs for Intelligence (Translation / Refinement)
google-genai>=0.1.0

# System utilities
argparse
tqdm
